Implementing wheel speed sensor for Arduino.

+++ arduino/leads_vec/WheelSpeedSensor.cpp
@@ -1,12 +1,12 @@
 #include "WheelSpeedSensor.h"
 #include "Algorithms.h"
 
-WheelSpeedSensor::WheelSpeedSensor(int *const pins) : Device<float>(pins) {
+WheelSpeedSensor::WheelSpeedSensor(int *const pins) : Device<int>(pins) {
 }
 
 void WheelSpeedSensor::initialize() { pinMode(_pins[0], INPUT); }
 
-float WheelSpeedSensor::read() {
+int WheelSpeedSensor::read() {
     if (millis() - _t2 > BOUNCETIME && pulseTriggered(_pins[0])) {
         _t1 = _t2;
         _t2 = millis();
